+++ src/app/components/search-components/brewery-list/brewery-list.constants.ts
@@ -1,10 +1,4 @@
 export const NO_ELEMENTS_FOUND = 'Sorry, no matching breweries found.';
-export const ENTER_SEARCH_TEXT = 'Enter text in search and press the search button, to search for breweries.';
 export const DEFAULT_SEARCH_LIMIT = 5;
 export const MAX_SEARCH_LIMIT = 10;
 
-export enum CARD_STATE {
-  SEARCH_HISTORY,
-  SEARCH_LIST,
-  BREWERY_DETAIL
-}
