+++ src/app/components/search-components/search-container/search-container.constants.ts
@@ -1,0 +1,8 @@
+export const ENTER_SEARCH_TEXT = 'Enter text in search and press the search button, to search for breweries.';
+
+export enum CARD_STATE {
+  NONE,
+  SEARCH_HISTORY,
+  SEARCH_LIST,
+  BREWERY_DETAIL
+}
